# backend/src/main/java/com/example/backend/dto/auth/request/SignInRequest.java
package com.example.backend.dto.auth.request;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.experimental.SuperBuilder;

@Setter
@Getter
@SuperBuilder
@NoArgsConstructor
@AllArgsConstructor
public class SignInRequest extends AuthRequest {
    @NotBlank(message = "ReCaptchaToken is required")
    private String reCaptchaToken;
}