# backend/src/main/java/com/example/backend/dto/receipt/ReceiptCreateDTO.java
package com.example.backend.dto.receipt;

import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;
import lombok.NoArgsConstructor;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ReceiptCreateDTO {
    private List<ReceiptRequestItemDTO> items;
}