# frontend/swagger.json
{
  "openapi": "3.1.0",
  "info": { "title": "API Documentation", "version": "1.0" },
  "servers": [
    { "url": "http://localhost:8080", "description": "Generated server url" }
  ],
  "security": [{ "Bearer Authentication": [] }],
...
  "paths": {
...
    "/api/products/{id}": {
      "get": {
        "tags": ["Product Management"],
        "summary": "Get a product",
        "description": "Fetch a product by ID.",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseResponseProductDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Product Management"],
        "summary": "Update a product",
        "description": "Update an existing product by ID.",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductRequestDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseResponseProductDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/receipts": {
      "get": {
        "tags": ["Receipt Controller"],
        "summary": "Get all receipts",
        "description": "Fetch a list of all registered receipts.",
        "operationId": "getReceipts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePaginateResponseReceiptResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Receipt Controller"],
        "summary": "Create a receipt",
        "description": "Create a receipt by owner or staff of the shop.",
        "operationId": "createReceipt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReceiptCreateDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseReceiptResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": ["Product Management"],
        "summary": "Get page products",
        "description": "Fetch a list of page registered products.",
        "operationId": "getProducts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10 }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePaginateResponseResponseProductDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Product Management"],
        "summary": "Create a product",
        "description": "Create a new product",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductRequestDTO" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseResponseProductDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/sign-up": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Sign up a new user",
        "description": "Sign up a new user with a username and password.",
        "operationId": "signUp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignUpRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseUserDTO" }
              }
            }
          }
        }
      }
    },
    "/api/auth/sign-in": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Sign in a user",
        "description": "Perform authentication on a user to sign in!.",
        "operationId": "signIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignInRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseSignInResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "resetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseVoid" }
              }
            }
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Forgot password",
        "description": "Validate email and send a code to reset password.",
        "operationId": "forgotPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ForgotPasswordRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/BaseResponseVoid" }
              }
            }
          }
        }
      }
    },
...
    "/api/receipts/{id}": {
      "get": {
        "tags": ["Receipt Controller"],
        "summary": "Get a receipt",
        "description": "Fetch a registered receipt.",
        "operationId": "getReceipt",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseReceiptResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseResponseUserDTO": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/UserDTO" },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "username": { "type": "string" },
          "name": { "type": "string" },
          "phoneNumber": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "role": { "type": "string" },
          "status": { "type": "string" },
          "shopId": { "type": "integer", "format": "int64" },
          "shopName": { "type": "string" },
          "email": { "type": "string" }
        },
        "required": [
          "createdAt",
          "email",
          "id",
          "name",
          "phoneNumber",
          "role",
          "status",
          "username"
        ]
      },
...
      "ProductRequestDTO": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "categoryId": { "type": "integer", "format": "int64" },
          "partnerId": { "type": "integer", "format": "int64" },
          "shopId": { "type": "integer", "format": "int64" },
          "imageUrls": { "type": "array", "items": { "type": "string" } }
        },
        "required": ["name", "shopId"]
      },
      "BaseResponseResponseProductDTO": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/ResponseProductDTO" },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "imageUrl": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE"] }
        },
        "required": ["status"]
      },
      "Partner": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "contactName": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "address": { "type": "string" },
          "website": { "type": "string" },
          "description": { "type": "string" }
        }
      },
      "ResponseProductDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "category": { "$ref": "#/components/schemas/Category" },
          "partner": { "$ref": "#/components/schemas/Partner" },
          "shopId": { "type": "integer", "format": "int64" },
          "shopName": { "type": "string" },
          "price": { "type": "number" },
          "imageUrls": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "ReceiptCreateDTO": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReceiptRequestItemDTO" }
          }
        }
      },
      "ReceiptRequestItemDTO": {
        "type": "object",
        "properties": {
          "productId": { "type": "integer", "format": "int64" },
          "quantity": { "type": "integer", "format": "int32", "minimum": 1 },
          "price": { "type": "number", "minimum": 0.0 },
          "zoneId": { "type": "integer", "format": "int64" }
        }
      },
      "BaseResponseReceiptResponseDTO": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/ReceiptResponseDTO" },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "ReceiptItemResponseDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "receiptId": { "type": "integer", "format": "int64" },
          "productId": { "type": "integer", "format": "int64" },
          "productName": { "type": "string" },
          "productPrice": { "type": "number", "format": "double" },
          "quantity": { "type": "integer", "format": "int32" },
          "zoneId": { "type": "integer", "format": "int64" },
          "zoneName": { "type": "string" }
        }
      },
      "ReceiptResponseDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "createdBy": { "$ref": "#/components/schemas/UserDTO" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "status": {
            "type": "string",
            "enum": ["SUCCESS", "FAILED", "PENDING"]
          },
          "receiptItems": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReceiptItemResponseDTO" }
          }
        }
      },
...
      "SignUpRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "maxLength": 20, "minLength": 3 },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 6
          },
          "phoneNumber": { "type": "string", "pattern": "^[0-9]{10,12}$" },
          "name": { "type": "string" },
          "email": { "type": "string" }
        },
        "required": ["email", "name", "password", "phoneNumber", "username"]
      },
      "SignInRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "maxLength": 20, "minLength": 3 },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 6
          },
          "reCaptchaToken": { "type": "string" }
        },
        "required": ["password", "reCaptchaToken", "username"]
      },
      "BaseResponseSignInResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/SignInResponse" },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "SignInResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/UserDTO" }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 6
          },
          "token": { "type": "string" }
        },
        "required": ["password", "token"]
      },
      "BaseResponseVoid": {
        "type": "object",
        "properties": {
          "data": { "type": "object" },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "reCaptchaToken": { "type": "string" }
        },
        "required": ["email", "reCaptchaToken"]
      },
...
      "BaseResponsePaginateResponseReceiptResponseDTO": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PaginateResponseReceiptResponseDTO"
          },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "PaginateResponseReceiptResponseDTO": {
        "type": "object",
        "properties": {
          "pageSize": { "type": "integer", "format": "int32" },
          "pageNumber": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int32" },
          "totalPages": { "type": "integer", "format": "int32" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReceiptResponseDTO" }
          }
        }
      },
      "BaseResponsePaginateResponseResponseProductDTO": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PaginateResponseResponseProductDTO"
          },
          "message": { "type": "string" },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_MISSING",
              "TOKEN_EXPIRED",
              "TOKEN_INVALID",
              "ACCESS_DENIED",
              "BAD_REQUEST",
              "INTERNAL_SERVER_ERROR"
            ]
          }
        }
      },
      "PaginateResponseResponseProductDTO": {
        "type": "object",
        "properties": {
          "pageSize": { "type": "integer", "format": "int32" },
          "pageNumber": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int32" },
          "totalPages": { "type": "integer", "format": "int32" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ResponseProductDTO" }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}